requirements.txt
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-mgmt-compute==33.0.0
azure-mgmt-core==1.5.0
azure-mgmt-storage==21.2.1
azure-monitor-query==1.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
fonttools==4.54.1
idna==3.10
isodate==0.7.2
kiwisolver==1.4.7
#matplotlib==3.9.2
#ymsal==1.31.0
#ymsal-extensions==1.2.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
ping3==4.0.8
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
